requirements.txt
flask==3.0.3
python-dotenv==1.0.1
supabase==2.4.6
requests==2.31.0
openpyxl==3.1.5
